[data] move many datasource classes for internal
