+++ python/ray/data/_internal/datasource/bigquery_datasource.py
@@ -4,12 +4,10 @@ from typing import List, Optional
 from ray.data._internal.util import _check_import
 from ray.data.block import Block, BlockMetadata
 from ray.data.datasource.datasource import Datasource, ReadTask
-from ray.util.annotations import PublicAPI
 
 logger = logging.getLogger(__name__)
 
 
-@PublicAPI(stability="alpha")
 class BigQueryDatasource(Datasource):
     def __init__(
         self,
