+++ python/ray/data/_internal/datasource/csv_datasource.py
@@ -2,13 +2,11 @@ from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Optional, Union
 
 from ray.data.block import Block
 from ray.data.datasource.file_based_datasource import FileBasedDatasource
-from ray.util.annotations import PublicAPI
 
 if TYPE_CHECKING:
     import pyarrow
 
 
-@PublicAPI
 class CSVDatasource(FileBasedDatasource):
     """CSV datasource, for reading and writing CSV files."""
