+++ python/ray/data/_internal/datasource/numpy_datasource.py
@@ -5,13 +5,11 @@ import numpy as np
 
 from ray.data.block import Block, BlockAccessor
 from ray.data.datasource.file_based_datasource import FileBasedDatasource
-from ray.util.annotations import PublicAPI
 
 if TYPE_CHECKING:
     import pyarrow
 
 
-@PublicAPI
 class NumpyDatasource(FileBasedDatasource):
     """Numpy datasource, for reading and writing Numpy files."""
