+++ python/ray/data/_internal/datasource/parquet_bulk_datasource.py
@@ -2,7 +2,6 @@ import logging
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
 from ray.data.datasource.file_based_datasource import FileBasedDatasource
-from ray.util.annotations import PublicAPI
 
 if TYPE_CHECKING:
     import pyarrow
@@ -11,7 +10,6 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-@PublicAPI
 class ParquetBulkDatasource(FileBasedDatasource):
     """Minimal Parquet datasource, for reading and writing Parquet files."""
