+++ python/ray/data/_internal/datasource/parquet_datasource.py
@@ -39,7 +39,6 @@ from ray.data.datasource.path_util import (
     _has_file_extension,
     _resolve_paths_and_filesystem,
 )
-from ray.util.annotations import PublicAPI
 
 if TYPE_CHECKING:
     import pyarrow
@@ -143,7 +142,6 @@ def _check_for_legacy_tensor_type(schema):
             )
 
 
-@PublicAPI
 class ParquetDatasource(Datasource):
     """Parquet datasource, for reading and writing Parquet files.
