+++ python/ray/data/_internal/datasource/range_datasource.py
@@ -9,10 +9,8 @@ from ray.data._internal.util import _check_pyarrow_version
 from ray.data.block import Block, BlockAccessor, BlockMetadata
 from ray.data.context import DataContext
 from ray.data.datasource import Datasource, ReadTask
-from ray.util.annotations import PublicAPI
 
 
-@PublicAPI
 class RangeDatasource(Datasource):
     """An example datasource that generates ranges of numbers from [0..n)."""
