+++ python/ray/data/_internal/datasource/webdataset_datasource.py
@@ -10,7 +10,6 @@ from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional, Union
 
 from ray.data.block import BlockAccessor
 from ray.data.datasource.file_based_datasource import FileBasedDatasource
-from ray.util.annotations import PublicAPI
 
 if TYPE_CHECKING:
     import pyarrow
@@ -300,7 +299,6 @@ def _make_iterable(block: BlockAccessor):
     return block.iter_rows(public_row_format=False)
 
 
-@PublicAPI(stability="alpha")
 class WebDatasetDatasource(FileBasedDatasource):
     """A Datasource for WebDataset datasets (tar format with naming conventions)."""
