+++ python/ray/data/datasource/__init__.py
@@ -1,8 +1,6 @@
+from ray.data._internal.datasource.sql_datasource import Connection
 from ray.data.datasource.bigquery_datasink import _BigQueryDatasink
-from ray.data.datasource.bigquery_datasource import BigQueryDatasource
-from ray.data.datasource.binary_datasource import BinaryDatasource
 from ray.data.datasource.csv_datasink import _CSVDatasink
-from ray.data.datasource.csv_datasource import CSVDatasource
 from ray.data.datasource.datasink import Datasink, DummyOutputDatasink
 from ray.data.datasource.datasource import (
     Datasource,
@@ -28,15 +26,10 @@ from ray.data.datasource.filename_provider import FilenameProvider
 from ray.data.datasource.image_datasink import _ImageDatasink
 from ray.data.datasource.image_datasource import ImageDatasource
 from ray.data.datasource.json_datasink import _JSONDatasink
-from ray.data.datasource.json_datasource import JSONDatasource
 from ray.data.datasource.lance_datasource import LanceDatasource
 from ray.data.datasource.mongo_datasink import _MongoDatasink
-from ray.data.datasource.mongo_datasource import MongoDatasource
 from ray.data.datasource.numpy_datasink import _NumpyDatasink
-from ray.data.datasource.numpy_datasource import NumpyDatasource
-from ray.data.datasource.parquet_bulk_datasource import ParquetBulkDatasource
 from ray.data.datasource.parquet_datasink import _ParquetDatasink
-from ray.data.datasource.parquet_datasource import ParquetDatasource
 from ray.data.datasource.parquet_meta_provider import ParquetMetadataProvider
 from ray.data.datasource.partitioning import (
     Partitioning,
@@ -44,31 +37,23 @@ from ray.data.datasource.partitioning import (
     PathPartitionFilter,
     PathPartitionParser,
 )
-from ray.data.datasource.range_datasource import RangeDatasource
 from ray.data.datasource.sql_datasink import _SQLDatasink
-from ray.data.datasource.sql_datasource import Connection, SQLDatasource
-from ray.data.datasource.text_datasource import TextDatasource
 from ray.data.datasource.tfrecords_datasink import _TFRecordDatasink
 from ray.data.datasource.torch_datasource import TorchDatasource
 from ray.data.datasource.webdataset_datasink import _WebDatasetDatasink
-from ray.data.datasource.webdataset_datasource import WebDatasetDatasource
 
 # Note: HuggingFaceDatasource should NOT be imported here, because
 # we want to only import the Hugging Face datasets library when we use
 # ray.data.from_huggingface() or HuggingFaceDatasource() directly.
 __all__ = [
     "BaseFileMetadataProvider",
-    "BinaryDatasource",
     "_BigQueryDatasink",
-    "BigQueryDatasource",
     "BlockBasedFileDatasink",
     "Connection",
     "_CSVDatasink",
-    "CSVDatasource",
     "Datasink",
     "Datasource",
     "_SQLDatasink",
-    "SQLDatasource",
     "DefaultFileMetadataProvider",
     "DummyOutputDatasink",
     "FastFileMetadataProvider",
@@ -78,29 +63,21 @@ __all__ = [
     "_ImageDatasink",
     "ImageDatasource",
     "_JSONDatasink",
-    "JSONDatasource",
     "LanceDatasource",
     "_NumpyDatasink",
-    "NumpyDatasource",
-    "ParquetBulkDatasource",
     "_ParquetDatasink",
-    "ParquetDatasource",
     "ParquetMetadataProvider",
     "PartitionStyle",
     "PathPartitionFilter",
     "PathPartitionParser",
     "Partitioning",
     "RandomIntRowDatasource",
-    "RangeDatasource",
     "_MongoDatasink",
-    "MongoDatasource",
     "ReadTask",
     "Reader",
     "RowBasedFileDatasink",
-    "TextDatasource",
     "_TFRecordDatasink",
     "TorchDatasource",
     "_WebDatasetDatasink",
-    "WebDatasetDatasource",
     "_S3FileSystemWrapper",
 ]
