+++ python/ray/data/datasource/mongo_datasink.py
@@ -1,12 +1,14 @@
 import logging
 from typing import Any, Iterable
 
+from ray.data._internal.datasource.mongo_datasource import (
+    _validate_database_collection_exist,
+)
 from ray.data._internal.delegating_block_builder import DelegatingBlockBuilder
 from ray.data._internal.execution.interfaces import TaskContext
 from ray.data._internal.util import _check_import
 from ray.data.block import Block, BlockAccessor
 from ray.data.datasource.datasink import Datasink
-from ray.data.datasource.mongo_datasource import _validate_database_collection_exist
 
 logger = logging.getLogger(__name__)
