+++ python/ray/data/datasource/parquet_meta_provider.py
@@ -12,7 +12,7 @@ from ray.util.annotations import DeveloperAPI
 if TYPE_CHECKING:
     import pyarrow
 
-    from ray.data.datasource.parquet_datasource import _SerializedFragment
+    from ray.data._internal.datasource.parquet_datasource import _SerializedFragment
 
 
 FRAGMENTS_PER_META_FETCH = 6
@@ -131,7 +131,7 @@ class ParquetMetadataProvider(FileMetadataProvider):
             must be returned in the same order as all input file fragments, such
             that `metadata[i]` always contains the metadata for `fragments[i]`.
         """
-        from ray.data.datasource.parquet_datasource import _SerializedFragment
+        from ray.data._internal.datasource.parquet_datasource import _SerializedFragment
 
         if len(fragments) > PARALLELIZE_META_FETCH_THRESHOLD:
             # Wrap Parquet fragments in serialization workaround.
@@ -167,7 +167,9 @@ def _fetch_metadata_serialization_wrapper(
     retry_max_attempts: int,
     retry_max_interval: int,
 ) -> List["pyarrow.parquet.FileMetaData"]:
-    from ray.data.datasource.parquet_datasource import _deserialize_fragments_with_retry
+    from ray.data._internal.datasource.parquet_datasource import (
+        _deserialize_fragments_with_retry,
+    )
 
     deserialized_fragments = _deserialize_fragments_with_retry(fragments)
     try:
@@ -187,24 +189,24 @@ def _fetch_metadata_serialization_wrapper(
             "To increase the maximum number of attempts, configure "
             "`RETRY_MAX_ATTEMPTS_FOR_META_FETCH_TASK`. For example:\n"
             "```\n"
-            "ray.data.datasource.parquet_datasource.RETRY_MAX_ATTEMPTS_FOR_META_FETCH_TASK = 64\n"  # noqa: E501
+            "ray.data._internal.datasource.parquet_datasource.RETRY_MAX_ATTEMPTS_FOR_META_FETCH_TASK = 64\n"  # noqa: E501
             "```\n"
             "To increase the maximum retry backoff interval, configure "
             "`RETRY_MAX_BACKOFF_S_FOR_META_FETCH_TASK`. For example:\n"
             "```\n"
-            "ray.data.datasource.parquet_datasource.RETRY_MAX_BACKOFF_S_FOR_META_FETCH_TASK = 128\n"  # noqa: E501
+            "ray.data._internal.datasource.parquet_datasource.RETRY_MAX_BACKOFF_S_FOR_META_FETCH_TASK = 128\n"  # noqa: E501
             "```\n"
             "If the error continues to occur, you can also try decresasing the "
             "concurency of metadata fetching tasks by setting "
             "`NUM_CPUS_FOR_META_FETCH_TASK` to a larger value. For example:\n"
             "```\n"
-            "ray.data.datasource.parquet_datasource.NUM_CPUS_FOR_META_FETCH_TASK = 4.\n"  # noqa: E501
+            "ray.data._internal.datasource.parquet_datasource.NUM_CPUS_FOR_META_FETCH_TASK = 4.\n"  # noqa: E501
             "```\n"
             "To change which exceptions to retry on, set "
             "`RETRY_EXCEPTIONS_FOR_META_FETCH_TASK` to a list of error messages. For "
             "example:\n"
             "```\n"
-            'ray.data.datasource.parquet_datasource.RETRY_EXCEPTIONS_FOR_META_FETCH_TASK = ["AWS Error ACCESS_DENIED", "Timeout"]\n'  # noqa: E501
+            'ray.data._internal.datasource.parquet_datasource.RETRY_EXCEPTIONS_FOR_META_FETCH_TASK = ["AWS Error ACCESS_DENIED", "Timeout"]\n'  # noqa: E501
             "```"
         ) from e
     return metadata
