+++ python/ray/data/datasource/sql_datasink.py
@@ -1,9 +1,9 @@
 from typing import Any, Callable, Iterable
 
+from ray.data._internal.datasource.sql_datasource import Connection, _connect
 from ray.data._internal.execution.interfaces import TaskContext
 from ray.data.block import Block, BlockAccessor
 from ray.data.datasource.datasink import Datasink
-from ray.data.datasource.sql_datasource import Connection, _connect
 
 
 class _SQLDatasink(Datasink):
