+++ python/ray/data/tests/test_optimize.py
@@ -8,9 +8,9 @@ import pytest
 import ray
 from ray._private.internal_api import memory_summary
 from ray.data import Dataset
+from ray.data._internal.datasource.csv_datasource import CSVDatasource
 from ray.data.block import BlockMetadata
 from ray.data.datasource import Datasource, ReadTask
-from ray.data.datasource.csv_datasource import CSVDatasource
 from ray.data.tests.util import column_udf, extract_values
 from ray.tests.conftest import *  # noqa
